requirements.txt
Flask==2.0.1
weasyprint==53.0
python-dateutil==2.8.2
requests==2.26.0
